djapp/requirements.txt
django==4.1
django-filter==23.5
django-extensions==3.2.3
djangorestframework==3.14.0
django-rest-registration==0.8.3
drf_yasg==1.21.7
gunicorn==21.2.0
httptools==0.6.1
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.25.0
uvloop==0.19.0
